Remove BfGraphQLGoogleAuth
Summary:

Since BfGoogleAuth isn't really a type on graphql, but instead it just augments other types, it's confusing to have it as its own type.

Test Plan:
