+++ packages/graphql/types/BfGraphQLPerson.ts
@@ -5,6 +5,10 @@ import {
 } from "packages/graphql/deps.ts";
 import { BfAccount } from "packages/bfDb/models/BfAccount.ts";
 import type { GraphQLContext } from "packages/graphql/graphql.ts";
+import { BfPerson } from "packages/bfDb/models/BfPerson.ts";
+import { getLogger } from "deps.ts";
+
+const logger = getLogger(import.meta);
 
 export const BfGraphQLPerson = objectType({
   name: "BfPerson",
@@ -13,6 +17,15 @@ export const BfGraphQLPerson = objectType({
     t.implements("BfNode");
     t.string("name");
     t.string("email");
+    t.string("googleAuthAccessToken", {
+      async resolve(parent, _, { bfCurrentViewer }) {
+        // @ts-expect-error #techdebt
+        const person = await BfPerson.findX(bfCurrentViewer, parent.id);
+        const auth = await person.getGoogleAuth();
+        logger.debug(`${person} - ${auth}`);
+        return await auth?.getAccessToken() ?? null;
+      },
+    });
     t.connectionField("accounts", {
       type: "BfAccount",
       async resolve(
